requirements.txt
Pillow==9.0.1
pip==20.0.2
pkg-resources==0.0.0
setuptools==44.0.0
tk==0.1.0
wheel==0.34.2
